+++ backend/test/de/tischner/cobweb/routing/model/graph/BasicGraphTest.java
@@ -18,7 +18,7 @@ public final class BasicGraphTest {
   /**
    * Counter used for generating unique edge IDs.
    */
-  private long mEdgeIdCounter;
+  private int mEdgeIdCounter;
   /**
    * The graph used for testing.
    */
@@ -30,12 +30,12 @@ public final class BasicGraphTest {
   @Before
   public void setUp() {
     mGraph = new BasicGraph();
-    final BasicNode firstNode = new BasicNode(1L);
-    final BasicNode secondNode = new BasicNode(2L);
-    final BasicNode thirdNode = new BasicNode(3L);
-    final BasicNode fourthNode = new BasicNode(4L);
-    final BasicNode fifthNode = new BasicNode(5L);
-    final BasicNode sixthNode = new BasicNode(6L);
+    final BasicNode firstNode = new BasicNode(1);
+    final BasicNode secondNode = new BasicNode(2);
+    final BasicNode thirdNode = new BasicNode(3);
+    final BasicNode fourthNode = new BasicNode(4);
+    final BasicNode fifthNode = new BasicNode(5);
+    final BasicNode sixthNode = new BasicNode(6);
 
     mGraph.addNode(firstNode);
     mGraph.addNode(secondNode);
@@ -61,7 +61,7 @@ public final class BasicGraphTest {
    */
   @Test
   public void testAddEdge() {
-    final BasicEdge<BasicNode> edge = new BasicEdge<>(10L, new BasicNode(10L), new BasicNode(11L), 1.0);
+    final BasicEdge<BasicNode> edge = new BasicEdge<>(10, new BasicNode(10), new BasicNode(11), 1.0);
     Assert.assertFalse(mGraph.containsEdge(edge));
     Assert.assertTrue(mGraph.addEdge(edge));
     Assert.assertTrue(mGraph.containsEdge(edge));
@@ -76,7 +76,7 @@ public final class BasicGraphTest {
   @Test
   public void testAddNode() {
     Assert.assertEquals(6, mGraph.size());
-    mGraph.addNode(new BasicNode(10L));
+    mGraph.addNode(new BasicNode(10));
     Assert.assertEquals(7, mGraph.size());
   }
 
@@ -100,7 +100,7 @@ public final class BasicGraphTest {
    */
   @Test
   public void testContainsEdge() {
-    final BasicEdge<BasicNode> edge = new BasicEdge<>(10L, new BasicNode(10L), new BasicNode(11L), 1.0);
+    final BasicEdge<BasicNode> edge = new BasicEdge<>(10, new BasicNode(10), new BasicNode(11), 1.0);
     Assert.assertFalse(mGraph.containsEdge(edge));
     mGraph.addEdge(edge);
     Assert.assertTrue(mGraph.containsEdge(edge));
@@ -108,14 +108,14 @@ public final class BasicGraphTest {
 
   /**
    * Test method for
-   * {@link de.tischner.cobweb.routing.model.graph.BasicGraph#containsNodeWithId(long)}.
+   * {@link de.tischner.cobweb.routing.model.graph.BasicGraph#containsNodeWithId(int)}.
    */
   @Test
   public void testContainsNodeWithId() {
-    Assert.assertTrue(mGraph.containsNodeWithId(1L));
-    Assert.assertTrue(mGraph.containsNodeWithId(5L));
-    Assert.assertFalse(mGraph.containsNodeWithId(8L));
-    Assert.assertFalse(mGraph.containsNodeWithId(-2L));
+    Assert.assertTrue(mGraph.containsNodeWithId(1));
+    Assert.assertTrue(mGraph.containsNodeWithId(5));
+    Assert.assertFalse(mGraph.containsNodeWithId(8));
+    Assert.assertFalse(mGraph.containsNodeWithId(-2));
   }
 
   /**
@@ -124,7 +124,7 @@ public final class BasicGraphTest {
    */
   @Test
   public void testGetAmountOfEdges() {
-    final BasicEdge<BasicNode> edge = new BasicEdge<>(10L, new BasicNode(10L), new BasicNode(11L), 1.0);
+    final BasicEdge<BasicNode> edge = new BasicEdge<>(10, new BasicNode(10), new BasicNode(11), 1.0);
     Assert.assertEquals(18, mGraph.getAmountOfEdges());
     mGraph.addEdge(edge);
     Assert.assertEquals(19, mGraph.getAmountOfEdges());
@@ -138,11 +138,11 @@ public final class BasicGraphTest {
    */
   @Test
   public void testGetEdges() {
-    final BasicNode first = new BasicNode(1L);
-    final BasicNode second = new BasicNode(2L);
-    final BasicNode third = new BasicNode(3L);
-    final BasicEdge<BasicNode> firstEdge = new BasicEdge<>(1L, first, second, 1.0);
-    final BasicEdge<BasicNode> secondEdge = new BasicEdge<>(2L, second, third, 1.0);
+    final BasicNode first = new BasicNode(1);
+    final BasicNode second = new BasicNode(2);
+    final BasicNode third = new BasicNode(3);
+    final BasicEdge<BasicNode> firstEdge = new BasicEdge<>(1, first, second, 1.0);
+    final BasicEdge<BasicNode> secondEdge = new BasicEdge<>(2, second, third, 1.0);
     mGraph = new BasicGraph();
     mGraph.addNode(first);
     mGraph.addNode(second);
@@ -153,10 +153,10 @@ public final class BasicGraphTest {
     Assert.assertTrue(mGraph.containsEdge(firstEdge));
     Assert.assertTrue(mGraph.containsEdge(secondEdge));
 
-    final Set<Long> destinationIds =
+    final Set<Integer> destinationIds =
         mGraph.getEdges().map(IEdge::getDestination).map(IHasId::getId).collect(Collectors.toSet());
-    Assert.assertTrue(destinationIds.contains(2L));
-    Assert.assertTrue(destinationIds.contains(3L));
+    Assert.assertTrue(destinationIds.contains(2));
+    Assert.assertTrue(destinationIds.contains(3));
   }
 
   /**
@@ -165,28 +165,28 @@ public final class BasicGraphTest {
    */
   @Test
   public void testGetIncomingEdges() {
-    final Set<Long> sourceIds = mGraph.getIncomingEdges(mGraph.getNodeById(1L).get()).stream().map(IEdge::getSource)
+    final Set<Integer> sourceIds = mGraph.getIncomingEdges(mGraph.getNodeById(1).get()).stream().map(IEdge::getSource)
         .map(IHasId::getId).collect(Collectors.toSet());
     Assert.assertEquals(4, sourceIds.size());
-    Assert.assertTrue(sourceIds.contains(2L));
-    Assert.assertTrue(sourceIds.contains(3L));
-    Assert.assertTrue(sourceIds.contains(4L));
-    Assert.assertTrue(sourceIds.contains(5L));
+    Assert.assertTrue(sourceIds.contains(2));
+    Assert.assertTrue(sourceIds.contains(3));
+    Assert.assertTrue(sourceIds.contains(4));
+    Assert.assertTrue(sourceIds.contains(5));
   }
 
   /**
    * Test method for
-   * {@link de.tischner.cobweb.routing.model.graph.BasicGraph#getNodeById(long)}.
+   * {@link de.tischner.cobweb.routing.model.graph.BasicGraph#getNodeById(int)}.
    */
   @Test
   public void testGetNodeById() {
-    Assert.assertTrue(mGraph.getNodeById(1L).isPresent());
-    Assert.assertEquals(1L, mGraph.getNodeById(1L).get().getId());
-    Assert.assertTrue(mGraph.getNodeById(5L).isPresent());
-    Assert.assertEquals(5L, mGraph.getNodeById(5L).get().getId());
+    Assert.assertTrue(mGraph.getNodeById(1).isPresent());
+    Assert.assertEquals(1L, mGraph.getNodeById(1).get().getId());
+    Assert.assertTrue(mGraph.getNodeById(5).isPresent());
+    Assert.assertEquals(5L, mGraph.getNodeById(5).get().getId());
 
-    Assert.assertFalse(mGraph.getNodeById(8L).isPresent());
-    Assert.assertFalse(mGraph.getNodeById(-2L).isPresent());
+    Assert.assertFalse(mGraph.getNodeById(8).isPresent());
+    Assert.assertFalse(mGraph.getNodeById(-2).isPresent());
   }
 
   /**
@@ -195,14 +195,14 @@ public final class BasicGraphTest {
    */
   @Test
   public void testGetNodes() {
-    final Set<Long> nodeIds = mGraph.getNodes().stream().map(IHasId::getId).collect(Collectors.toSet());
+    final Set<Integer> nodeIds = mGraph.getNodes().stream().map(IHasId::getId).collect(Collectors.toSet());
     Assert.assertEquals(6, nodeIds.size());
-    Assert.assertTrue(nodeIds.contains(1L));
-    Assert.assertTrue(nodeIds.contains(2L));
-    Assert.assertTrue(nodeIds.contains(3L));
-    Assert.assertTrue(nodeIds.contains(4L));
-    Assert.assertTrue(nodeIds.contains(5L));
-    Assert.assertTrue(nodeIds.contains(6L));
+    Assert.assertTrue(nodeIds.contains(1));
+    Assert.assertTrue(nodeIds.contains(2));
+    Assert.assertTrue(nodeIds.contains(3));
+    Assert.assertTrue(nodeIds.contains(4));
+    Assert.assertTrue(nodeIds.contains(5));
+    Assert.assertTrue(nodeIds.contains(6));
 
     Assert.assertTrue(new BasicGraph().getNodes().isEmpty());
   }
@@ -213,13 +213,13 @@ public final class BasicGraphTest {
    */
   @Test
   public void testGetOutgoingEdges() {
-    final Set<Long> destinationIds = mGraph.getOutgoingEdges(mGraph.getNodeById(1L).get()).stream()
+    final Set<Integer> destinationIds = mGraph.getOutgoingEdges(mGraph.getNodeById(1).get()).stream()
         .map(IEdge::getDestination).map(IHasId::getId).collect(Collectors.toSet());
     Assert.assertEquals(4, destinationIds.size());
-    Assert.assertTrue(destinationIds.contains(2L));
-    Assert.assertTrue(destinationIds.contains(3L));
-    Assert.assertTrue(destinationIds.contains(4L));
-    Assert.assertTrue(destinationIds.contains(5L));
+    Assert.assertTrue(destinationIds.contains(2));
+    Assert.assertTrue(destinationIds.contains(3));
+    Assert.assertTrue(destinationIds.contains(4));
+    Assert.assertTrue(destinationIds.contains(5));
   }
 
   /**
@@ -228,7 +228,7 @@ public final class BasicGraphTest {
    */
   @Test
   public void testRemoveEdge() {
-    final BasicEdge<BasicNode> edge = new BasicEdge<>(10L, new BasicNode(10L), new BasicNode(11L), 1.0);
+    final BasicEdge<BasicNode> edge = new BasicEdge<>(10, new BasicNode(10), new BasicNode(11), 1.0);
     mGraph.addEdge(edge);
     Assert.assertTrue(mGraph.containsEdge(edge));
     Assert.assertTrue(mGraph.removeEdge(edge));
@@ -243,15 +243,15 @@ public final class BasicGraphTest {
    */
   @Test
   public void testRemoveNode() {
-    final BasicNode node = new BasicNode(10L);
+    final BasicNode node = new BasicNode(10);
     Assert.assertEquals(6, mGraph.size());
     Assert.assertFalse(mGraph.removeNode(node));
     Assert.assertEquals(6, mGraph.size());
     mGraph.addNode(node);
     Assert.assertEquals(7, mGraph.size());
-    Assert.assertTrue(mGraph.containsNodeWithId(10L));
+    Assert.assertTrue(mGraph.containsNodeWithId(10));
     Assert.assertTrue(mGraph.removeNode(node));
-    Assert.assertFalse(mGraph.containsNodeWithId(10L));
+    Assert.assertFalse(mGraph.containsNodeWithId(10));
     Assert.assertEquals(6, mGraph.size());
   }
 
@@ -261,11 +261,11 @@ public final class BasicGraphTest {
    */
   @Test
   public void testReverse() {
-    final BasicNode first = new BasicNode(1L);
-    final BasicNode second = new BasicNode(2L);
-    final BasicNode third = new BasicNode(3L);
-    final BasicEdge<BasicNode> firstEdge = new BasicEdge<>(1L, first, second, 1.0);
-    final BasicEdge<BasicNode> secondEdge = new BasicEdge<>(2L, second, third, 1.0);
+    final BasicNode first = new BasicNode(1);
+    final BasicNode second = new BasicNode(2);
+    final BasicNode third = new BasicNode(3);
+    final BasicEdge<BasicNode> firstEdge = new BasicEdge<>(1, first, second, 1.0);
+    final BasicEdge<BasicNode> secondEdge = new BasicEdge<>(2, second, third, 1.0);
     mGraph = new BasicGraph();
     mGraph.addNode(first);
     mGraph.addNode(second);
@@ -278,10 +278,10 @@ public final class BasicGraphTest {
     Assert.assertTrue(mGraph.getEdges().anyMatch(edge -> firstEdge.equals(edge)));
     Assert.assertTrue(mGraph.getEdges().anyMatch(edge -> secondEdge.equals(edge)));
 
-    final Set<Long> destinationIds =
+    final Set<Integer> destinationIds =
         mGraph.getEdges().map(IEdge::getDestination).map(IHasId::getId).collect(Collectors.toSet());
-    Assert.assertTrue(destinationIds.contains(1L));
-    Assert.assertTrue(destinationIds.contains(2L));
+    Assert.assertTrue(destinationIds.contains(1));
+    Assert.assertTrue(destinationIds.contains(2));
   }
 
   /**
@@ -291,7 +291,7 @@ public final class BasicGraphTest {
   @Test
   public void testSize() {
     Assert.assertEquals(6, mGraph.size());
-    mGraph.addNode(new BasicNode(10L));
+    mGraph.addNode(new BasicNode(10));
     Assert.assertEquals(7, mGraph.size());
 
     Assert.assertEquals(0, new BasicGraph().size());
